Щас всё исправно, BackUp point

+++ Assets/Scripts/Obstacle.cs
@@ -7,10 +7,12 @@ public class Obstacle : MonoBehaviour
     private void OnCollisionEnter(Collision other)
     {
         Ally ally = other.gameObject.GetComponent<Ally>();
-        if (ally != null)
+        if (ally != null && AlliesGroup.Instance.Count != 0)
         {
+            
             AlliesGroup.Instance.Kill(ally);
-            AlliesGroup.Instance.UpdateRingsValue();
+            
+           
         }
     }
 }
